observe: la rÃ©solution change en fonction de la fenÃªtre !
donc il faut mettre Ã  jour le fichier quand on reÃ§oit le message, mais
n'afficher le bandeau que la 1Ãšre fois

+++ observe.lua
@@ -1,11 +1,20 @@
 local mp = require 'mp'
 
 function on_width_change(name,value)
+	local file = io.open("video_size","r")
+	local exists = false
+	if file then
+		io.close(file)
+		exists = true
+	end
 	height = mp.get_property("osd-height")
 	if (value > 0 and tonumber(height) > 0) then
 		file = io.open("video_size", "w")
 		file:write(string.format("%d\n%d\n",value,height))
 		file:close()
+	end
+	if exists then
+		return
 	end
 	os.remove("list_coords")
 	os.remove("info_coords")
